+++ linux/runtime/proxy-off.sh
@@ -1,5 +1,5 @@
 # shellcheck shell=sh
 # Disable account proxy variables in the current shell.
 
-unset CLASH_SSH_PROXY CLASH_SSH_NO_PROXY CODEX_CLASH_PROXY
+unset CLASH_SSH_PROXY CLASH_SSH_NO_PROXY CODEX_CLASH_PROXY NODE_USE_ENV_PROXY
 unset http_proxy https_proxy HTTP_PROXY HTTPS_PROXY ALL_PROXY all_proxy NO_PROXY no_proxy